requirements.txt
requests>=2.31.0
anthropic>=0.34.0
facebook-sdk>=3.1.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
apscheduler>=3.10.4
Pillow>=10.0.0
lxml>=4.9.0

